# src/docs/swagger.yaml
definitions:
...
  enum.CollectionItemType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    type: integer
    x-enum-varnames:
    - StringType
    - IntType
    - FloatType
    - BoolType
    - DateTimeType
    - EnumType
    - NestedCollectionType
    - OtherEntityType
    - NoType
  enum.CollectionType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - List
    - Set
    - Map
    - Array
    - None
  enum.DataType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    type: integer
    x-enum-varnames:
    - String
    - Int
    - Float
    - Bool
    - DateTime
    - Enum
    - Collection
    - Entity
  enum.DbInteraction:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - ORM
    - DBFunction
    - RawSQL
    - NoSql
    - NoDB
  enum.DerivativeType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - Arithmetic
    - Formula
    - Concatenation
    - Runtime
    - NotDerived
  enum.DisplayStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - Summary
    - Detail
    - Hide
    - Show
  enum.PreferredDB:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    type: integer
    x-enum-varnames:
    - Postgres
    - Mysql
    - Mongo
    - Mssql
    - Maria
    - Oracle
    - NA
  enum.ProgrammingLanguage:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    type: integer
    x-enum-varnames:
    - Golang
    - Python
    - Java
    - JavaScript
    - TypeScript
    - Csharp
    - Cpp
    - Kotlin
    - Rust
  enum.ProjectType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - Enterprise
    - Website
  enum.RelationType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - OneToOne
    - OneToMany
    - ManyToOne
    - ManyToMany
    - NoRelation
...
  gen-concept-api_api_dto.DependsOnEntity:
    properties:
      entityName:
        type: string
      fieldName:
        type: string
      relationType:
        $ref: '#/definitions/enum.RelationType'
    type: object
  gen-concept-api_api_dto.Entity:
    properties:
      dependsOnEntities:
        items:
          $ref: '#/definitions/gen-concept-api_api_dto.DependsOnEntity'
        type: array
      entityDescription:
        type: string
      entityFields:
        items:
          $ref: '#/definitions/gen-concept-api_api_dto.EntityField'
        type: array
      entityName:
        type: string
      implementsAudit:
        type: boolean
      implementsChangeManagement:
        type: boolean
      implementsRBAC:
        type: boolean
      isAuthenticationRequired:
        type: boolean
      isBackendOnly:
        type: boolean
      isIndependentEntity:
        type: boolean
      isReadOnly:
        type: boolean
      modeOfDBInteraction:
        $ref: '#/definitions/enum.DbInteraction'
      preferredDB:
        $ref: '#/definitions/enum.PreferredDB'
      version:
        type: string
    type: object
  gen-concept-api_api_dto.EntityField:
    properties:
      collectionEntity:
        type: string
      collectionItemType:
        $ref: '#/definitions/enum.CollectionItemType'
      collectionType:
        $ref: '#/definitions/enum.CollectionType'
      derivativeExpression:
        type: string
      derivativeType:
        $ref: '#/definitions/enum.DerivativeType'
      displayName:
        type: string
      displayStatus:
        $ref: '#/definitions/enum.DisplayStatus'
      enumValues:
        items:
          type: string
        type: array
      fieldDescription:
        type: string
      fieldName:
        type: string
      fieldType:
        $ref: '#/definitions/enum.DataType'
      inputValidations:
        items:
          $ref: '#/definitions/gen-concept-api_api_dto.InputValidation'
        type: array
      isBackendOnly:
        type: boolean
      isCollection:
        type: boolean
      isDerived:
        type: boolean
      isEditable:
        type: boolean
      isEnum:
        type: boolean
      isMandatory:
        type: boolean
      isReadOnly:
        type: boolean
      isSensitive:
        type: boolean
      isUnique:
        type: boolean
      nestedCollectionItemType:
        $ref: '#/definitions/enum.CollectionItemType'
      sampleData:
        type: string
    type: object
  gen-concept-api_api_dto.FileResponse:
    properties:
      description:
        type: string
      directory:
        type: string
      id:
        type: integer
      mimeType:
        type: string
      name:
        type: string
    type: object
...
  gen-concept-api_api_dto.InputValidation:
    properties:
      abortOnFailure:
        type: boolean
      customErrorMessage:
        type: string
      description:
        type: string
    type: object
...
  gen-concept-api_api_dto.Project:
    properties:
      entities:
        items:
          $ref: '#/definitions/gen-concept-api_api_dto.Entity'
        type: array
      isMultiLingual:
        type: boolean
      isMultiTenant:
        type: boolean
      programmingLanguage:
        $ref: '#/definitions/enum.ProgrammingLanguage'
      projectDescription:
        type: string
      projectName:
        type: string
      projectType:
        $ref: '#/definitions/enum.ProjectType'
    type: object
...
  gen-concept-api_api_dto.UpdateFileRequest:
    properties:
      description:
        type: string
    type: object
...
  gen-concept-api_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/gen-concept-api_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/gen-concept-api_api_validation.ValidationError'
        type: array
    type: object
  gen-concept-api_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  gen-concept-api_api_validation.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
  gen-concept-api_domain_filter.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description: contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
  gen-concept-api_domain_filter.PagedList-gen-concept-api_api_dto_FileResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/gen-concept-api_api_dto.FileResponse'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  gen-concept-api_domain_filter.PagedList-gen-concept-api_api_dto_Project:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/gen-concept-api_api_dto.Project'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  gen-concept-api_domain_filter.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/gen-concept-api_domain_filter.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/gen-concept-api_domain_filter.Sort'
        type: array
    type: object
  gen-concept-api_domain_filter.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/files/:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a file
      parameters:
      - in: formData
        name: description
        required: true
        type: string
      - description: Create a file
        in: formData
        name: file
        required: true
        type: file
      responses:
        "201":
          description: File response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_api_dto.FileResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a file
      tags:
      - Files
  /v1/files/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a file
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a file
      tags:
      - Files
    get:
      consumes:
      - application/json
      description: Get a file
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: File response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_api_dto.FileResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a file
      tags:
      - Files
    put:
      consumes:
      - application/json
      description: Update a file
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a file
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/gen-concept-api_api_dto.UpdateFileRequest'
      responses:
        "200":
          description: File response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_api_dto.FileResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a file
      tags:
      - Files
  /v1/files/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Files
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/gen-concept-api_domain_filter.PaginationInputWithFilter'
      responses:
        "200":
          description: File response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_domain_filter.PagedList-gen-concept-api_api_dto_FileResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Files
      tags:
      - Files
...
  /v1/projects/:
    post:
      consumes:
      - application/json
      description: Create a Project
      parameters:
      - description: Create a Project
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/gen-concept-api_api_dto.Project'
      responses:
        "201":
          description: Project response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_api_dto.Project'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a Project
      tags:
      - Projects
  /v1/projects/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Project
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a Project
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: Get a Project
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Project response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_api_dto.Project'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a Project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Update a Project
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a Project
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/gen-concept-api_api_dto.Project'
      responses:
        "200":
          description: Project response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_api_dto.Project'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a Project
      tags:
      - Projects
  /v1/projects/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Properties
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/gen-concept-api_domain_filter.PaginationInputWithFilter'
      responses:
        "200":
          description: Project response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_domain_filter.PagedList-gen-concept-api_api_dto_Project'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Properties
      tags:
      - Projects
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"